server/requirements.txt
langchain==0.3.27
langchain-openai==0.3.35
langgraph==0.6.8
python-dotenv==1.1.1
fastapi==0.118.0
uvicorn==0.37.0